refactor: improve theme cubit
