+++ lib/app/service_locator/modules/theme_cubit_module.dart
@@ -1,6 +1,6 @@
 import 'package:get_it/get_it.dart';
 
-import '../../../features/theme/theme_cubit.dart';
+import '../../../core/shared/cubits/theme_cubit/theme_cubit.dart';
 
 void registerThemeCubit(GetIt sl) {
   sl.registerLazySingleton(() => ThemeCubit());
